+++ src/drafts/src/drafts/transaction_draft.cpp
@@ -87,6 +87,16 @@ namespace drafts
     PositionId TradeLegDraft::getPositionId() const { return _positionId; }
 
     /**
+     * @brief Get the currency associated with the trade leg draft.
+     *
+     * @return Currency The currency associated with the trade leg draft.
+     */
+    Currency TradeLegDraft::getCurrency() const
+    {
+        return _unitPrice.getCurrency();
+    }
+
+    /**
      * @brief Set the instrument ID associated with the trade leg draft.
      *
      * @param instrumentId The instrument ID to associate with the trade leg
@@ -161,6 +171,19 @@ namespace drafts
     }
 
     /**
+     * @brief Adds multiple entries to the cash transaction draft.
+     *
+     * @param entries The vector of TransactionEntryDraft to add to the cash
+     * transaction draft.
+     */
+    void CreateTransactionDraft::addEntries(
+        const std::vector<TransactionEntryDraft>& entries
+    )
+    {
+        _entries.insert(_entries.end(), entries.begin(), entries.end());
+    }
+
+    /**
      * @brief Create a Stock Transaction Draft:: Create Stock Transaction Draft
      * object
      *
@@ -213,12 +236,19 @@ namespace drafts
      *
      * @param accountId
      * @param cash
+     * @param type
+     * @param isExternal
      */
     TransactionEntryDraft::TransactionEntryDraft(
-        AccountId     accountId,
-        finance::Cash cash
+        AccountId            accountId,
+        finance::Cash        cash,
+        TransactionEntryType type,
+        bool                 isExternal
     )
-        : _accountId(accountId), _cash(cash)
+        : _accountId(accountId),
+          _cash(cash),
+          _isExternal(isExternal),
+          _type(type)
     {
     }
 
@@ -260,6 +290,35 @@ namespace drafts
      * @return bool
      */
     bool TransactionEntryDraft::needsExternal() const { return _needsExternal; }
+
+    /**
+     * @brief get whether this transaction entry draft is for an external
+     * account
+     *
+     * @return bool
+     */
+    bool TransactionEntryDraft::isExternal() const { return _isExternal; }
+
+    /**
+     * @brief get the type of this transaction entry draft
+     *
+     * @return TransactionEntryType
+     */
+    TransactionEntryType TransactionEntryDraft::getType() const
+    {
+        return _type;
+    }
+
+    /**
+     * @brief Get the currency associated with this transaction entry draft.
+     *
+     * @return Currency The currency associated with this transaction entry
+     * draft.
+     */
+    Currency TransactionEntryDraft::getCurrency() const
+    {
+        return _cash.getCurrency();
+    }
 
     /**
      * @brief Construct a new Transaction Overview Draft:: Transaction Overview
@@ -339,6 +398,96 @@ namespace drafts
     ) const
     {
         return _comment;
+    }
+
+    /**
+     * @brief get the total general cash of the transaction overview draft
+     *
+     * @return finance::Cash
+     */
+    finance::Cash TransactionOverviewDraft::getTotalGeneralCash() const
+    {
+        auto totalGeneralCash = finance::Cash(getCurrency());
+        for (const auto& entry : _entries)
+        {
+            if (entry.getType() == TransactionEntryType::General &&
+                !entry.isExternal())
+            {
+                totalGeneralCash += entry.getCash();
+            }
+        }
+        return totalGeneralCash;
+    }
+
+    /**
+     * @brief get the total fees of the transaction overview draft
+     *
+     * @return finance::Cash
+     */
+    finance::Cash TransactionOverviewDraft::getTotalFees() const
+    {
+        auto totalFees = finance::Cash(getCurrency());
+        for (const auto& entry : _entries)
+        {
+            if (entry.getType() == TransactionEntryType::Fees &&
+                !entry.isExternal())
+            {
+                totalFees += entry.getCash();
+            }
+        }
+        return totalFees;
+    }
+
+    /**
+     * @brief Get the currency associated with this transaction overview draft.
+     *
+     * @return Currency The currency associated with this transaction overview
+     * draft.
+     */
+    Currency TransactionOverviewDraft::getCurrency() const
+    {
+        if (!_entries.empty())
+            return _entries.front().getCurrency();
+
+        if (!_legs.empty())
+            return _legs.front().getCurrency();
+
+        return Currency();
+    }
+
+    /**
+     * @brief Get the account ID associated with the first leg of the
+     * transaction overview draft.
+     *
+     * @return AccountId The account ID associated with the first leg of the
+     * transaction overview draft.
+     */
+    AccountId TransactionOverviewDraft::getLegAccount() const
+    {
+        if (!_legs.empty())
+            return _legs.front().getAccountId();
+
+        return AccountId::invalid();
+    }
+
+    /**
+     * @brief Get the account ID associated with the first entry of the
+     * transaction overview draft.
+     *
+     * @param includeExternal Whether to include external entries.
+     *
+     * @return AccountId The account ID associated with the first entry of the
+     * transaction overview draft.
+     */
+    AccountId TransactionOverviewDraft::getEntryAccountId(
+        bool includeExternal
+    ) const
+    {
+        for (const auto& entry : _entries)
+            if (!entry.isExternal() || includeExternal)
+                return entry.getAccountId();
+
+        return AccountId::invalid();
     }
 
 }   // namespace drafts