+++ src/ui/include/ui/transaction/amount_row.hpp
@@ -30,9 +30,11 @@ namespace ui
         explicit AmountRow(QWidget* parent = nullptr);
 
         void setNDecimalPlaces(std::uint8_t places);
+        void setDefaultValue(int value);
 
         [[nodiscard]] bool        isValid() const;
-        [[nodiscard]] micro_units getAmount() const;
+        [[nodiscard]] bool        isZero() const;
+        [[nodiscard]] micro_units getAmount(std::size_t precision) const;
 
        signals:
         /**