+++ src/ui/src/ui/transaction/stock_widget.cpp
@@ -5,7 +5,10 @@
 #include <qformlayout.h>
 #include <qlabel.h>
 #include <qpushbutton.h>
+#include <qwidget.h>
 
+#include "config/finance.hpp"
+#include "drafts/account_draft.hpp"
 #include "drafts/transaction_draft.hpp"
 #include "finance/cash.hpp"
 #include "finance/currency.hpp"
@@ -14,6 +17,7 @@
 #include "ui/transaction/comment_field.hpp"
 #include "ui/transaction/ticker_field.hpp"
 #include "ui/transaction/timestamp_field.hpp"
+#include "ui/utils/error.hpp"
 #include "utils/qt_helpers.hpp"
 
 using utils::makeQChild;
@@ -22,56 +26,261 @@ namespace ui
 {
 
     /**
+     * @brief Struct holding the fields for the stock widget
+     *
+     */
+    struct StockWidget::Fields
+    {
+        /// The combo box for selecting the primary account
+        AccountCombo* accountCombo;
+
+        /// The combo box for selecting the reference account
+        AccountCombo* referenceAccountCombo;
+
+        /// The row for entering the quantity of the stock
+        AmountRow* quantityRow;
+
+        /// The row for entering the price of the stock
+        AmountRow* priceRow;
+
+        /// The row for entering the fees of the stock
+        AmountRow* feesRow;
+
+        /// The label for displaying the currency of the selected account
+        QLabel* currencyLabel;
+
+        /// The field for entering the stock ticker
+        TickerField* tickerField;
+
+        /// The field for entering the timestamp of the transaction
+        TimestampField* timestampField;
+
+        /// The field for entering a comment about the transaction
+        CommentField* commentField;
+
+        /// The label for displaying the currency of the fees
+        QLabel* currencyFeesLabel;
+
+        Fields(
+            const std::vector<drafts::AccountDraft>& accounts,
+            const std::vector<drafts::AccountDraft>& referenceAccounts,
+            const std::vector<std::string>&          tickers,
+            QWidget*                                 parent
+        );
+
+        void addFieldsToLayout(QFormLayout* layout) const;
+
+        [[nodiscard]]
+        bool isValid() const;
+
+        void update() const;
+
+        [[nodiscard]]
+        drafts::CreateStockTransactionDraft getDraft() const;
+    };
+
+    /**
+     * @brief Construct a new Stock Widget:: Fields:: Fields object
+     *
+     * @param accounts
+     * @param referenceAccounts
+     * @param tickers
+     * @param parent
+     */
+    StockWidget::Fields::Fields(
+        const std::vector<drafts::AccountDraft>& accounts,
+        const std::vector<drafts::AccountDraft>& referenceAccounts,
+        const std::vector<std::string>&          tickers,
+        QWidget*                                 parent
+    )
+        : accountCombo(makeQChild<AccountCombo>(accounts, parent)),
+          referenceAccountCombo(
+              makeQChild<AccountCombo>(referenceAccounts, parent)
+          ),
+          quantityRow(makeQChild<AmountRow>(parent)),
+          priceRow(makeQChild<AmountRow>(parent)),
+          feesRow(makeQChild<AmountRow>(parent)),
+          currencyLabel(makeQChild<QLabel>(parent)),
+          tickerField(makeQChild<TickerField>(tickers, parent)),
+          timestampField(makeQChild<TimestampField>(parent)),
+          commentField(makeQChild<CommentField>(parent)),
+          currencyFeesLabel(makeQChild<QLabel>(parent))
+    {
+        feesRow->setDefaultValue(0);
+    }
+
+    /**
+     * @brief Add the fields to the layout
+     *
+     * @param layout The layout to add the fields to
+     */
+    void StockWidget::Fields::addFieldsToLayout(QFormLayout* layout) const
+    {
+        layout->addRow("Account:", accountCombo);
+        layout->addRow("Reference Account:", referenceAccountCombo);
+        layout->addRow("Ticker:", tickerField);
+        layout->addRow("Timestamp:", timestampField);
+
+        auto* quantityRowLayout = makeQChild<QHBoxLayout>();
+        quantityRowLayout->addWidget(quantityRow);
+        layout->addRow("Quantity:", quantityRowLayout);
+        quantityRow->setNDecimalPlaces(Quantity::precision);
+
+        auto* amountRowLayout = makeQChild<QHBoxLayout>();
+        amountRowLayout->addWidget(priceRow);
+        amountRowLayout->addWidget(currencyLabel);
+        layout->addRow("Stock Price:", amountRowLayout);
+        auto* feesRowLayout = makeQChild<QHBoxLayout>();
+        feesRowLayout->addWidget(feesRow);
+        feesRowLayout->addWidget(currencyFeesLabel);
+        layout->addRow("Fees:", feesRowLayout);
+        layout->addRow("Comment:", commentField);
+
+        referenceAccountCombo->setEnabled(false);
+    }
+
+    /**
+     * @brief Check if the fields are valid
+     *
+     * @return true if all fields are valid, false otherwise
+     */
+    bool StockWidget::Fields::isValid() const
+    {
+        if (!accountCombo->selected().has_value())
+            return false;
+        if (!referenceAccountCombo->selected().has_value())
+            return false;
+
+        if (!tickerField->isValid())
+            return false;
+        if (!quantityRow->isValid())
+            return false;
+        if (!priceRow->isValid())
+            return false;
+
+        if (quantityRow->isZero())
+            return false;
+
+        if (priceRow->isZero())
+            return false;
+
+        return true;
+    }
+
+    /**
+     * @brief Update the fields
+     *
+     */
+    void StockWidget::Fields::update() const
+    {
+        accountCombo->update();
+        referenceAccountCombo->update();
+        quantityRow->update();
+        priceRow->update();
+        feesRow->update();
+        currencyLabel->update();
+        tickerField->update();
+        currencyFeesLabel->update();
+    }
+
+    /**
+     * @brief Get the draft for the stock transaction
+     *
+     * @return drafts::CreateStockTransactionDraft
+     */
+    drafts::CreateStockTransactionDraft StockWidget::Fields::getDraft() const
+    {
+        const auto account = accountCombo->selected();
+
+        if (!account.has_value() || account.value().id == AccountId::invalid())
+            throw std::runtime_error("No account selected");
+
+        const auto referenceAccount = referenceAccountCombo->selected();
+
+        if (!referenceAccount.has_value() ||
+            referenceAccount.value().id == AccountId::invalid())
+            throw std::runtime_error("No reference account selected");
+
+        const auto currency     = referenceAccount->currency;
+        const auto refAccountId = referenceAccount->id;
+        const auto microUnits   = finance::getMicroUnit(currency);
+
+        const auto unitPrice_ = priceRow->getAmount(microUnits);
+        const auto unitPrice  = finance::Cash(currency, unitPrice_);
+        const auto quantity_  = quantityRow->getAmount(Quantity::precision);
+        const auto quantity   = Quantity{quantity_};
+        const auto cash       = -quantity * unitPrice;
+
+        auto entry = drafts::TransactionEntryDraft{
+            refAccountId,
+            cash,
+            TransactionEntryType::General,
+            false
+        };
+
+        const auto fees_ = feesRow->getAmount(microUnits);
+        const auto fees  = -finance::Cash(currency, fees_);
+
+        const auto feesEntry = drafts::TransactionEntryDraft{
+            refAccountId,
+            fees,
+            TransactionEntryType::Fees,
+            false
+        };
+
+        const auto ticker = tickerField->getTicker();
+        if (!ticker.has_value())
+            throw std::runtime_error("No ticker selected");
+
+        auto tradeLeg = drafts::TradeLegDraft{
+            account->id,
+            unitPrice,
+            quantity,
+            ticker.value()
+        };
+
+        return {
+            timestampField->getTimestamp(),
+            {entry, feesEntry},
+            {tradeLeg},
+            commentField->getComment()
+        };
+    }
+
+    /**
      * @brief Construct a new Stock Widget:: Stock Widget object
      *
-     * @param accounts A list of account drafts to populate the account combo
-     * box
+     * @param accounts A list of account drafts to populate the account
+     * combo box
      * @param referenceAccounts A list of account drafts to populate the
-     * reference account combo box, this will be filtered based on the selected
-     * primary account to only include accounts with the same currency
+     * reference account combo box, this will be filtered based on the
+     * selected primary account to only include accounts with the same
+     * currency
      * @param tickers A list of ticker symbols to populate the ticker field
      * @param parent The parent widget for this widget
      */
     StockWidget::StockWidget(
-        std::vector<drafts::AccountDraft>        accounts,
+        const std::vector<drafts::AccountDraft>& accounts,
         const std::vector<drafts::AccountDraft>& referenceAccounts,
         const std::vector<std::string>&          tickers,
         QWidget*                                 parent
     )
         : Dialog(parent),
           _layout(new QFormLayout(this)),
-          _accountCombo(makeQChild<AccountCombo>(std::move(accounts), this)),
-          _referenceAccountCombo(
-              makeQChild<AccountCombo>(referenceAccounts, this)
-          ),
-          _quantityRow(makeQChild<AmountRow>(this)),
-          _priceRow(makeQChild<AmountRow>(this)),
-          _currencyLabel(makeQChild<QLabel>(this)),
           _addButton(makeQChild<QPushButton>("Add Transaction", this)),
-          _tickerField(makeQChild<TickerField>(tickers, this)),
-          _timestampField(makeQChild<TimestampField>(this)),
-          _commentField(makeQChild<CommentField>(this)),
+          _fields(
+              std::make_unique<Fields>(
+                  accounts,
+                  referenceAccounts,
+                  tickers,
+                  this
+              )
+          ),
           _referenceAccounts(referenceAccounts)
     {
         setLayout(_layout);
 
-        _layout->addRow("Account:", _accountCombo);
-        _layout->addRow("Reference Account:", _referenceAccountCombo);
-        _layout->addRow("Ticker:", _tickerField);
-        _layout->addRow("Timestamp:", _timestampField);
-
-        auto* quantityRowLayout = makeQChild<QHBoxLayout>();
-        quantityRowLayout->addWidget(_quantityRow);
-        _layout->addRow("Quantity:", quantityRowLayout);
-        _quantityRow->setNDecimalPlaces(Quantity::precision);
-
-        auto* amountRowLayout = makeQChild<QHBoxLayout>();
-        amountRowLayout->addWidget(_priceRow);
-        amountRowLayout->addWidget(_currencyLabel);
-        _layout->addRow("Stock Price:", amountRowLayout);
-        _layout->addRow("Comment:", _commentField);
-
-        _referenceAccountCombo->setEnabled(false);
+        _fields->addFieldsToLayout(_layout);
 
         _addButton->setEnabled(false);
         auto* buttonLayout = makeQChild<QHBoxLayout>();
@@ -79,20 +288,20 @@ namespace ui
         _layout->addRow(buttonLayout);
 
         connect(
-            _accountCombo,
+            _fields->accountCombo,
             &AccountCombo::accountSelected,
             this,
             &StockWidget::_onAccountSelected
         );
         connect(
-            _referenceAccountCombo,
+            _fields->referenceAccountCombo,
             &AccountCombo::accountSelected,
             this,
             &StockWidget::_onReferenceAccountSelected
         );
         _connectAddButton();
         connect(
-            _tickerField,
+            _fields->tickerField,
             &TickerField::createTickerRequested,
             this,
             &StockWidget::createTickerRequested
@@ -100,6 +309,8 @@ namespace ui
 
         connect(_addButton, &QPushButton::clicked, this, &StockWidget::_emitOk);
     }
+
+    StockWidget::~StockWidget() = default;
 
     /**
      * @brief Handle the selection of a primary account, this will be called
@@ -120,19 +331,24 @@ namespace ui
         using finance::getMicroUnit;
         using finance::getSymbol;
 
-        _priceRow->setNDecimalPlaces(getMicroUnit(account.currency));
-        _currencyLabel->setText(getSymbol(account.currency).c_str());
+        const auto currency   = account.currency;
+        const auto microUnits = getMicroUnit(currency);
+
+        _fields->priceRow->setNDecimalPlaces(microUnits);
+        _fields->feesRow->setNDecimalPlaces(microUnits);
+        _fields->currencyLabel->setText(getSymbol(currency).c_str());
+        _fields->currencyFeesLabel->setText(getSymbol(currency).c_str());
 
         std::vector<drafts::AccountDraft> referenceAccounts;
 
         for (const auto& referenceAccount : _referenceAccounts)
         {
-            if (referenceAccount.currency == account.currency)
+            if (referenceAccount.currency == currency)
                 referenceAccounts.push_back(referenceAccount);
         }
 
-        _referenceAccountCombo->updateAccounts(std::move(referenceAccounts));
-        _referenceAccountCombo->setEnabled(true);
+        _fields->referenceAccountCombo->updateAccounts(referenceAccounts);
+        _fields->referenceAccountCombo->setEnabled(true);
         _updateAddButton();
     }
 
@@ -167,12 +383,7 @@ namespace ui
      */
     void StockWidget::_updateAddButton()
     {
-        const auto isValid = _accountCombo->selected().has_value() &&
-                             _referenceAccountCombo->selected().has_value() &&
-                             _priceRow->isValid() &&
-                             _quantityRow->getAmount() != 0 &&
-                             _priceRow->getAmount() != 0 &&
-                             _quantityRow->isValid() && _tickerField->isValid();
+        const auto isValid = _fields->isValid();
 
         _addButton->setEnabled(isValid);
     }
@@ -189,7 +400,15 @@ namespace ui
      */
     void StockWidget::_emitOk()
     {
-        emit createStockTransactionRequested(_getDraft());
+        try
+        {
+            const auto draft = _getDraft();
+            emit       createStockTransactionRequested(draft);
+        }
+        catch (const std::exception& e)
+        {
+            ErrorDialog::show(QString(e.what()));
+        }
     }
 
     /**
@@ -199,7 +418,7 @@ namespace ui
      */
     void StockWidget::updateAccounts(std::vector<drafts::AccountDraft> accounts)
     {
-        _accountCombo->updateAccounts(std::move(accounts));
+        _fields->accountCombo->updateAccounts(std::move(accounts));
     }
 
     /**
@@ -215,7 +434,7 @@ namespace ui
     {
         _referenceAccounts = std::move(referenceAccounts);
 
-        const auto account = _accountCombo->selected();
+        const auto account = _fields->accountCombo->selected();
         if (account.has_value())
             _onAccountSelected(account.value());
     }
@@ -233,23 +452,14 @@ namespace ui
         for (const auto& ticker : tickers)
             qTickers.emplace_back(QString::fromStdString(ticker));
 
-        _tickerField->updateTickers(std::move(qTickers));
+        _fields->tickerField->updateTickers(std::move(qTickers));
     }
 
     /**
      * @brief Refresh the widget to reflect the current state
      *
      */
-    void StockWidget::refresh()
-    {
-        _accountCombo->update();
-        _referenceAccountCombo->update();
-        _quantityRow->update();
-        _priceRow->update();
-        _currencyLabel->update();
-        _addButton->update();
-        _tickerField->update();
-    }
+    void StockWidget::refresh() { _fields->update(); }
 
     /**
      * @brief Gets the draft for the stock transaction.
@@ -259,43 +469,7 @@ namespace ui
      */
     drafts::CreateStockTransactionDraft StockWidget::_getDraft() const
     {
-        const auto account = _accountCombo->selected();
-
-        if (!account.has_value() || account.value().id == AccountId::invalid())
-            throw std::runtime_error("No account selected");
-
-        const auto referenceAccount = _referenceAccountCombo->selected();
-
-        if (!referenceAccount.has_value() ||
-            referenceAccount.value().id == AccountId::invalid())
-            throw std::runtime_error("No reference account selected");
-
-        const auto unitPrice =
-            finance::Cash(referenceAccount->currency, _priceRow->getAmount());
-
-        const auto quantity = Quantity{_quantityRow->getAmount()};
-
-        const auto cash = -quantity * unitPrice;
-
-        auto entry = drafts::TransactionEntryDraft{referenceAccount->id, cash};
-
-        const auto ticker = _tickerField->getTicker();
-        if (!ticker.has_value())
-            throw std::runtime_error("No ticker selected");
-
-        auto tradeLeg = drafts::TradeLegDraft{
-            account->id,
-            unitPrice,
-            quantity,
-            ticker.value()
-        };
-
-        return {
-            _timestampField->getTimestamp(),
-            {entry},
-            {tradeLeg},
-            _commentField->getComment()
-        };
+        return _fields->getDraft();
     }
 
     /**
@@ -305,19 +479,19 @@ namespace ui
     void StockWidget::_connectAddButton()
     {
         connect(
-            _priceRow,
+            _fields->priceRow,
             &AmountRow::validityChanged,
             this,
             &StockWidget::_updateAddButton
         );
         connect(
-            _priceRow,
+            _fields->priceRow,
             &AmountRow::valueChanged,
             this,
             &StockWidget::_updateAddButton
         );
         connect(
-            _tickerField,
+            _fields->tickerField,
             &TickerField::tickerSelected,
             this,
             &StockWidget::_updateAddButton